requirements.txt
Flask
SQLAlchemy
psycopg2-binary
Flask-SQLAlchemy
pytest
pytest-cov
coverage
python-dotenv
